+++ WhistleWindLobotomyMod/Abilities/Special Abilities/CensoredSpecial.cs
@@ -12,7 +12,7 @@ namespace WhistleWindLobotomyMod {
         public SpecialTriggeredAbility SpecialAbility => specialAbility;
 
         public const string rName = "CENSORED";
-        public const string rDesc = "Whenver CENSORED kills a card, create a CENSORED in your hand with the killed card's Power, tribes, and sigils.";
+        public const string rDesc = "Whenver CENSORED kills a card, create a CENSORED in your hand with the killed card's Power.";
 
         public override bool RespondsToOtherCardDie(PlayableCard card, CardSlot deathSlot, bool fromCombat, PlayableCard killer) {
             if (fromCombat)
@@ -25,48 +25,17 @@ namespace WhistleWindLobotomyMod {
             // Creates a minion that has the abilities, tribes, power of the killed card
             CardInfo minion = CardLoader.GetCardByName(Cards.censoredMinion);
 
-            minion.displayedName = card.Info.displayedName;
-            minion.appearanceBehaviour = card.Info.appearanceBehaviour;
-            minion.cost = card.Info.BloodCost;
-            minion.bonesCost = card.Info.BonesCost;
-            minion.energyCost = card.Info.EnergyCost;
-            minion.gemsCost = card.Info.GemsCost;
+            int newAttack = card.Info.baseAttack > 0 ? card.Info.baseAttack - 1 : 0;
 
-            int newAttack = card.Info.baseAttack < 1 ? 1 : card.Info.baseAttack;
-
-            minion.Mods.Add(new(newAttack, 1));
-
-            // Add tribes
-            foreach (Tribe item in card.Info.tribes.FindAll((Tribe x) => x != Tribe.NUM_TRIBES))
-                minion.tribes.Add(item);
-
-            // Add base sigils
-            foreach (Ability item in card.Info.abilities.FindAll((Ability x) => x != Ability.NUM_ABILITIES))
-                minion.Mods.Add(new CardModificationInfo(item));
-
-            foreach (CardModificationInfo item in card.Info.Mods.FindAll((CardModificationInfo x) => !x.nonCopyable)) {
-                // Add merged sigils
-                CardModificationInfo cardModificationInfo = (CardModificationInfo)item.Clone();
-                cardModificationInfo.healthAdjustment = 0;
-                minion.Mods.Add(cardModificationInfo);
-            }
+            minion.Mods.Add(new(newAttack, 0) {
+                nameReplacement = card.Info.displayedName
+            });
 
             base.PlayableCard.Anim.StrongNegationEffect();
             yield return new WaitForSeconds(0.4f);
 
             // create minion in hand if not an opponent, otherwise add to queue
-            if (!base.PlayableCard.OpponentCard) {
-                if (Singleton<ViewManager>.Instance.CurrentView != View.Hand) {
-                    yield return new WaitForSeconds(0.2f);
-                    Singleton<ViewManager>.Instance.SwitchToView(View.Hand, false, false);
-                    yield return new WaitForSeconds(0.2f);
-                }
-                yield return Singleton<CardSpawner>.Instance.SpawnCardToHand(minion);
-                yield return new WaitForSeconds(0.45f);
-            }
-            else {
-                CombatHelpers.QueueCreatedCard(minion);
-            }
+            yield return CombatHelpers.QueueOrCreateDrawnCard(minion, base.PlayableCard.OpponentCard);
             yield return DialogueHelper.PlayDialogueEvent("CENSOREDKilledCard");
             yield return new WaitForSeconds(0.25f);
         }