+++ WhistleWindLobotomyMod/CHANGELOG.md
@@ -54,10 +54,11 @@
 - <span style='color:#2084A0'>Beauty and the Beast: changed play cost from 1 Blood -> 1 Bone
 - Big and Will Be Bad Wolf: reworked to 2/1; Blood Lust; 2 Blood
 - <span style='color:#DE6080'>Binah: Reduced Health from 5 -> 3
-- <span style='color:#2084A0'>CENSORED: Reworked to 4/1; Bloodfiend; 2 Blood
+- CENSORED: Reworked to 4/1; Bloodfiend; 2 Blood
+- CENSORED: cards created by special ability now only inherit killed cards' Power, cost 2 Bones
 - Don't Touch Me: updated artwork
 - Don't Touch Me: no longer belongs to the Mechanical tribe
-- <span style='color:#2084A0'>Dream of a Black Swan: Reworked from 2/5 -> 4/4
+- Dream of a Black Swan: Reworked from 2/5 -> 4/4
 - Der Freischütz: reworked from 2/2, Sniper, Persistent, 2 Blood -> 1/2, Gun For Hire, Persistent, 1 Blood
 - Funeral of the Dead Butterflies: rebalanced from 1/3; 2 Blood -> 1/1; 1 Blood
 - Funeral of the Dead Butterflies: now belongs to the Undead temple (Act 2)